separate userauth node
